Added cleanup timer
Added cleanup timer that runs every 5 minutes and deletes all records that have expired.
The expiration date is set to 1 hour after the record was created.

docker-compose now also uses a Dockerfile to build the image.
This was necessary to install cron and copy the cleanup script to the image.

Finally, init-db.sh was modified so ha expirationdate now has the correct datatype.